Added left() and right() methods and free functions to Vector2D. left() transforms <x, y> into <-y, x> right() transforms <x, y > into <y, -x>
Added unit tests for left() and right()

Added BS_BUILD_DEBUG to debug builds of the UnitTests Visual Studio project

+++ tests/src/test/Bounds2.cpp
@@ -302,7 +302,7 @@ UT_TEST_BEGIN( Bounds2_setDimension )
     Bounds2i o( cv_values );
 
     for( size_t i = 0; i < cv_size; ++i )
-        o.setDimension( i, 10 + i );
+        o.setDimension( i, (int)(10 + i) );
 
     return allEqual( o.data, cv_dimTest );
 UT_TEST_END()
